Move out logic in util and put simple test case to test the new selection
